pyproject.toml
[project]
name = "raindrop-client"
version = "0.1.0"
description = "Python client library for the Raindrop.io API"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "respx>=0.22.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/raindrop_client"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
